AddAzureTableStores - adding Role support only when .AddRoles<>() is called prior in pipeline. Minor test cleanup
