+++ src/ElCamino.AspNetCore.Identity.AzureTable/IdentityAzureTableBuilderExtensions.cs
@@ -9,24 +9,35 @@ namespace Microsoft.Extensions.DependencyInjection
 {
     public static class IdentityAzureTableBuilderExtensions
     {
+        /// <summary>
+        /// Call .AddRoles<IdentityRole>() in the pipeline if you need Roles functionality, otherwise the RoleStore will not be loaded.
+        /// </summary>
+        /// <typeparam name="TContext"></typeparam>
+        /// <param name="builder"></param>
+        /// <param name="configAction"></param>
+        /// <returns></returns>
         public static IdentityBuilder AddAzureTableStores<TContext>(this IdentityBuilder builder, Func<IdentityConfiguration> configAction)
             where TContext : IdentityCloudContext, new()
         {
             builder.Services.AddSingleton<IdentityConfiguration>(new Func<IServiceProvider, IdentityConfiguration>(p => configAction()));
 
             Type contextType = typeof(TContext);
-            Type userStoreType = typeof(UserStore<,,>).MakeGenericType(builder.UserType, builder.RoleType ?? typeof(ElCamino.AspNetCore.Identity.AzureTable.Model.IdentityRole), contextType);
-            Type roleStoreType = typeof(RoleStore<,>).MakeGenericType(builder.RoleType??typeof(ElCamino.AspNetCore.Identity.AzureTable.Model.IdentityRole), contextType);
-
             builder.Services.AddScoped(contextType, contextType);
+
+            Type userStoreType = builder.RoleType != null ? typeof(UserStore<,,>).MakeGenericType(builder.UserType, builder.RoleType, contextType)
+                : typeof(UserOnlyStore<,>).MakeGenericType(builder.UserType, contextType);
 
             builder.Services.AddScoped(
                 typeof(IUserStore<>).MakeGenericType(builder.UserType),
                 userStoreType);
-            builder.Services.AddScoped(
-                typeof(IRoleStore<>).MakeGenericType(builder.RoleType ?? typeof(ElCamino.AspNetCore.Identity.AzureTable.Model.IdentityRole)),
-                roleStoreType);
 
+            if (builder.RoleType != null)
+            {
+                Type roleStoreType = typeof(RoleStore<,>).MakeGenericType(builder.RoleType, contextType);
+
+                builder.Services.AddScoped(
+                    typeof(IRoleStore<>).MakeGenericType(builder.RoleType), roleStoreType);
+            }
             return builder;
         }
